+++ src/icons/index.js
@@ -1,0 +1,6 @@
+import InfoIcon from './InfoIcon'
+import SuccessIcon from './SuccessIcon'
+import ErrorIcon from './ErrorIcon'
+import CloseIcon from './CloseIcon'
+
+export { InfoIcon, SuccessIcon, ErrorIcon, CloseIcon, BaseIcon }
